Update isAttested state after sending the transaction
- Merge `useIsAttested` hook into `useAttest` to simplify state
  management
- Update `isAttested` after sending the transaction.
